pcn-firewall: introduce datapath pipeline early break
this commit introduces the support for some firewall
filtering pipeline modules to early break the pipeline
if no rule is matched.

the idea is that during the execution of the pipeline
some modules are implemented in a way in wich the cost
of checking if the resulting bitvector is all zero
is almost negligible.

this support enables to speedup the execution since is no
longer needed to go through all the pipeline modules
if this condition is matched.

Signed-off-by: Matteo Bertrone <m.bertrone@gmail.com>